Extract type into its own file
