+++ Sources/Player/ResourceLoading/ResourceLoadedPlayerItem.swift
@@ -1,0 +1,23 @@
+//
+//  Copyright (c) SRG SSR. All rights reserved.
+//
+//  License information is available from the LICENSE file.
+//
+
+import AVFoundation
+
+private let kResourceLoaderQueue = DispatchQueue(label: "ch.srgssr.player.resource_loader")
+
+/// An item which stores its own custom resource loader delegate.
+final class ResourceLoadedPlayerItem: AVPlayerItem {
+    // swiftlint:disable:next weak_delegate
+    private let resourceLoaderDelegate: AVAssetResourceLoaderDelegate
+
+    init(url: URL, resourceLoaderDelegate: AVAssetResourceLoaderDelegate) {
+        self.resourceLoaderDelegate = resourceLoaderDelegate
+        let asset = AVURLAsset(url: url)
+        asset.resourceLoader.setDelegate(resourceLoaderDelegate, queue: kResourceLoaderQueue)
+        // Provide same key as for a standard asset, see `AVPlayerItem.init(asset:)` documentation.
+        super.init(asset: asset, automaticallyLoadedAssetKeys: ["duration"])
+    }
+}
